Add badges, break app into components
