+++ src/components/SettingsModal.tsx
@@ -1,12 +1,13 @@
 import { useRef } from "react";
-import Modal from "react-bootstrap/Modal";
-import Form from "react-bootstrap/Form";
-import "./Settings.css";
-import { FeVibeCheckConfig } from "../../src-tauri/bindings/FeVibeCheckConfig";
 import Button from "react-bootstrap/Button";
 import InputGroup from "react-bootstrap/InputGroup";
 import { invoke } from "@tauri-apps/api";
+import Modal from "react-bootstrap/Modal";
+import Form from "react-bootstrap/Form";
+
+import { FeVibeCheckConfig } from "../../src-tauri/bindings/FeVibeCheckConfig";
 import { SET_CONFIG } from "../data/constants";
+import "./SettingsModal.css";
 
 type settingsModalProps = {
   settings: FeVibeCheckConfig | null; // Nullable because backend sets default, don't want to overwrite if no response
@@ -15,40 +16,21 @@ type settingsModalProps = {
   onSave: () => void;
 };
 
-// I think we need to do something like this:
-// type InvokeableConfig = FeVibeCheckConfig & { [key: string]: unknown };
-// Because InvokeArgs has type [key:string]:unkown, we need to union w/ the interface to be able to pass into invoke()
-
-// Right now FeOSCNetworking is {bind:string, remote:string}
-// So we also need to be careful to not modify the other attributes
-
-// The signature of set_vibecheck_config is:
-//    (vc_state: tauri::State<'_, vcore::VCStateMutex>, fe_vc_config: FeVibeCheckConfig) -> Result<(), frontend::VCFeError>
-// So we need to also wrap the InvokeableType with feVcConfig
-type InvokeableConfig = { feVcConfig: FeVibeCheckConfig } & {
-  [key: string]: unknown;
-};
-
 export default function (props: settingsModalProps) {
   const oscBind = useRef<HTMLInputElement>(null);
 
   async function setConfig() {
-    if (
-      oscBind.current == null ||
-      oscBind.current.value == null ||
-      props.settings == null
-    ) {
+    if (oscBind.current?.value == null || props.settings == null) {
       return;
     }
-    const newConfig: InvokeableConfig = {
-      feVcConfig: {
-        networking: {
-          bind: oscBind.current.value,
-          remote: props.settings.networking.remote,
-        },
+    const newConfig: FeVibeCheckConfig = {
+      ...props.settings,
+      networking: {
+        ...props.settings.networking,
+        bind: oscBind.current.value,
       },
     };
-    await invoke(SET_CONFIG, newConfig);
+    await invoke(SET_CONFIG, { feVcConfig: newConfig });
   }
 
   if (props.settings == null) {
