+++ src/utils/index.ts
@@ -4,7 +4,7 @@ export const percent = new Intl.NumberFormat("en-US", {
   maximumFractionDigits: 0,
 });
 
-export const round2 = new Intl.NumberFormat("en-US", {
-  minimumFractionDigits: 2,
-  maximumFractionDigits: 2,
+export const round0 = new Intl.NumberFormat("en-US", {
+  minimumFractionDigits: 0,
+  maximumFractionDigits: 0,
 });
